@@ -1,15 +1,14 @@
 package au.com.eventsecretary.code;
 
-import au.com.eventsecretary.common.Identifiable;
 import au.com.eventsecretary.common.codes.CodeContext;
 import au.com.eventsecretary.persistence.BusinessObjectPersistence;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
 import java.util.List;
 
-@Configuration
+@AutoConfiguration
 public class CodeConfiguration {
     @Bean
     public CodeService codeService(BusinessObjectPersistence persistence) {
