@@ -1,9 +1,10 @@
 package au.com.eventsecretary.render;
 
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-@Configuration
+@AutoConfiguration
 public class RenderConfiguration {
     @Bean
     public Renderer renderer(freemarker.template.Configuration freemarkerConfiguration) {
